♻️ Refactor: separate fetch and combine vidoe and place data
