+++ src/components/MyRouteContent/index.tsx
@@ -18,6 +18,7 @@ import instance from '@/src/api/axios';
 import { useRouter } from 'next/router';
 import { useFilteredData } from '@/src/hooks/useFilteredData';
 import { CardDataItem } from '@/src/lib/types';
+import combineVideoPlace from '@/src/api/combineVideoPlace';
 // import { useRelatedSearch } from '@/src/hooks/useRelatedSearch';
 // import RelatedSearchInfo from '../mainContent/ListSearchSection/RelatedSearchInfo';
 
@@ -53,44 +54,10 @@ const MyRouteContent = () => {
 
   // fetch my place data and use its video id to fetch course data and combine them to create new data
   useEffect(() => {
-    const fetchMyPlace = async () => {
-      try {
-        if (!userId) return;
-        const videoData = await instance.get(`/user/${userId}/video`);
-        const modifiedVideoData = videoData.data.data.map((item) => ({
-          content: item.content,
-          id: item.id,
-          videoUrl: item.videoUrl,
-          tags: item.tags,
-          title: item.title,
-        }));
-
-        // Promise.all to fetch each course data then map them to combine with video data
-        const videoId = videoData.data.data.map((item) => item.id);
-        Promise.all(videoId.map((id) => instance.get(`/course/${id}`))).then((responses) => {
-          // responses is an array of responses for each request
-          const combinedData = responses.map((response, index) => {
-            const courseData = response.data.data.course[0];
-            return {
-              content: modifiedVideoData[index].content,
-              id: modifiedVideoData[index].id,
-              videoUrl: modifiedVideoData[index].videoUrl,
-              tags: modifiedVideoData[index].tags,
-              title: modifiedVideoData[index].title,
-              name: courseData.name,
-              img: courseData.img.replace(/'/g, ''),
-              description: courseData.description,
-              posX: courseData.posX,
-              posY: courseData.posY,
-            };
-          });
-          setMyPlaceData(combinedData);
-        });
-      } catch (error) {
-        console.error('Error fetching card list:', error);
-      }
+    const fetchData = async () => {
+      await combineVideoPlace(userId).then((data) => setMyPlaceData(data));
     };
-    fetchMyPlace();
+    fetchData();
   }, [setMyPlaceData, userId]);
 
   // call when dnd ends
